requirements.txt
asgiref==3.2.10
boto==2.49.0
colorama==0.4.4
colorlog==4.6.2
Django==3.0.8
django-ckeditor==5.9.0
django-constance==2.7.0
django-js-asset==1.2.2
django-markwhat==1.6.2
django-picklefield==3.0.1
gunicorn==19.9.0
keyring==5.7.1
markdown2==2.3.10
Pillow
python-memcached==1.59
pytz==2020.1
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
tornado==6.1
urllib3==1.26.2
whitenoise==5.2.0
django-environ==0.4.4
django-heroku
dj-database-url
psycopg2
